+++ code_backend/shared_config.py
@@ -24,9 +24,12 @@ from dotenv import load_dotenv
 from dataclasses import dataclass
 
 
-frontend_window_size: list = [800, 400]
-json_path: str = 'Databases/JSON_Files/spotify_devices.json'
-no_image_path: str = 'Icons/Spotify_if_no_image.png'
+# local paths for files
+json_path: str = '/home/simon/git_repos/MusicPlayer/Databases/JSON_Files/spotify_devices.json'
+no_image_path: str = '/home/simon/git_repos/MusicPlayer/Icons/Spotify_if_no_image.png'
+main_database_path = '/home/simon/git_repos/MusicPlayer/Databases/main_database.db'
+track_analysis_tsv_file_path = "/home/simon/git_repos/MusicPlayer/Databases/CSV_Files/track_analysis.tsv"
+
 max_popularity = 20
 skipping_steps: dict = {
     "20sec": -2,
@@ -35,6 +38,8 @@ skipping_steps: dict = {
     "last_quarter": 0,
     "not_skipped": 1
 }
+market = 'DE'
+frontend_window_size: list = [800, 400]
 
 
 # noinspection DuplicatedCode